Caught bugs and missing fields
Some of the fields weren't "spitting out" what I expected , investigated and corrected. For example declined_rep had not been formatted at all. already_represented had a lot of blank values when should've been only TRUE or FALSE.  
I also attempted to reformat the 8-digit A#s as leading with a zero digit but no avail. Tried a paste0 as.character solution and a formatC solution.

Lots of ideas here https://stackoverflow.com/questions/5812493/how-to-add-leading-zeros can revisit if time allows but not a huge deal.

**Next steps would be to try and create plots and summaries that can be used in reporting out on the due date.***